Changing package name from model to models.
