+++ src/main/java/uk/gov/companieshouse/acsp/controllers/AcspController.java
@@ -5,7 +5,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import uk.gov.companieshouse.acsp.exception.ServiceException;
-import uk.gov.companieshouse.acsp.model.dto.AcspDataDto;
+import uk.gov.companieshouse.acsp.models.dto.AcspDataDto;
 import uk.gov.companieshouse.acsp.service.AcspService;
 import uk.gov.companieshouse.acsp.service.TransactionService;
 import uk.gov.companieshouse.logging.Logger;
