+++ src/main/java/uk/gov/companieshouse/acsp/mapper/ACSPRegDataDtoDaoMapper.java
@@ -2,8 +2,8 @@ package uk.gov.companieshouse.acsp.mapper;
 
 import org.mapstruct.Mapper;
 import org.springframework.stereotype.Component;
-import uk.gov.companieshouse.acsp.model.dao.AcspDataDao;
-import uk.gov.companieshouse.acsp.model.dto.AcspDataDto;
+import uk.gov.companieshouse.acsp.models.dao.AcspDataDao;
+import uk.gov.companieshouse.acsp.models.dto.AcspDataDto;
 
 @Component
 @Mapper(componentModel = "spring")
