+++ src/main/java/uk/gov/companieshouse/acsp/repositories/AcspRepository.java
@@ -2,7 +2,7 @@ package uk.gov.companieshouse.acsp.repositories;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
-import uk.gov.companieshouse.acsp.model.dao.AcspDataDao;
+import uk.gov.companieshouse.acsp.models.dao.AcspDataDao;
 
 @Repository
 //For now datatype is String, Change the datatype later accordingly
