+++ src/main/java/uk/gov/companieshouse/acsp/service/AcspService.java
@@ -5,9 +5,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import uk.gov.companieshouse.acsp.exception.ServiceException;
 import uk.gov.companieshouse.acsp.mapper.ACSPRegDataDtoDaoMapper;
-import uk.gov.companieshouse.acsp.model.dao.AcspDataDao;
-import uk.gov.companieshouse.acsp.model.dao.AcspDataSubmissionDao;
-import uk.gov.companieshouse.acsp.model.dto.AcspDataDto;
+import uk.gov.companieshouse.acsp.models.dao.AcspDataDao;
+import uk.gov.companieshouse.acsp.models.dao.AcspDataSubmissionDao;
+import uk.gov.companieshouse.acsp.models.dto.AcspDataDto;
 import uk.gov.companieshouse.acsp.repositories.AcspRepository;
 import uk.gov.companieshouse.acsp.util.ApiLogger;
 import uk.gov.companieshouse.api.model.transaction.Resource;
