+++ src/test/java/uk/gov/companieshouse/acsp/service/AcspServiceTest.java
@@ -7,8 +7,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import uk.gov.companieshouse.acsp.model.dao.AcspDataDao;
-import uk.gov.companieshouse.acsp.model.dto.AcspDataDto;
+import uk.gov.companieshouse.acsp.models.dao.AcspDataDao;
+import uk.gov.companieshouse.acsp.models.dto.AcspDataDto;
 import uk.gov.companieshouse.acsp.repositories.AcspRepository;
 import uk.gov.companieshouse.api.model.transaction.Transaction;
 import uk.gov.companieshouse.acsp.mapper.ACSPRegDataDtoDaoMapper;
